+++ src/main/kotlin/dev/vitorvidal/inventory/api/domain/vo/product/RegisterProductVO.kt
@@ -3,5 +3,5 @@ package dev.vitorvidal.inventory.api.domain.vo.product
 data class RegisterProductVO(
     val productName: String,
     val productBrand: String,
-    val productPrice: Long
+    val productPrice: Double
 )
